requirements.txt
# PigeonPhoto_Automator - Requirements
# Python 3.9+

# Core Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0

# Segmentation
rembg>=2.0.50
onnxruntime>=1.16.0

# Pose Estimation
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
scipy>=1.11.0

# Image Processing
Pillow>=10.0.0

# Web Framework (MVP)
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0

